Split dashboard into multiple views.
