+++ src/components/Dashboard/AccountRequests.js
@@ -2,34 +2,56 @@ import React, { Component } from "react";
 import { connect } from "react-redux";
 import Container from "react-bootstrap/Container";
 import Form from "react-bootstrap/Form";
+import Card from "react-bootstrap/Card";
+import Col from "react-bootstrap/Col";
+import Table from "react-bootstrap/Table";
 import "./Dashboard.css";
 
 class UpdateAccountPendingStatus extends Component {
-  componentDidMount() {
-    this.props.dispatch({ type: 'FETCH_PENDING_ACCOUNTS' });
+  state = {
+    pendingInterval: null
   }
 
+  componentDidMount = () => {
+    const pendingInterval = setInterval(
+      () => this.props.dispatch({ type: "FETCH_PENDING_ACCOUNTS" }),
+      10 * 1000
+    );
+    this.setState({
+      pendingInterval: pendingInterval,
+    });
+  };
+
+  componentWillUnmount = () => {
+    clearInterval(this.state.pendingInterval);
+  };
+
   render() {
-    console.log(this.props.pendingAccounts);
     return (
       <>
         <Container>
           <Form.Row xs={12}>
-            <>
-              cool
-              {this.props.pendingAccounts.map((account, index) =>
-                <button
-                  key={`button-to-accept-account-${index}`}
-                  onChange={() => this.props.dispatch({ type: 'UPDATE_PENDING_ACCOUNT', payload: account })}
-                />
-              )}
-              <button
-                // key={`button-to-accept-account-${index}`}
-                onChange={() =>
-                  this.props.dispatch({ type: 'UPDATE_PENDING_ACCOUNT', payload: 'account' })
-                }
-              />
-            </>
+            <Col xs={12} sm={12} md={12} lg={4} xl={4}>
+              <Card>
+                <form className="dashForm">
+                  <h1 id="thirdColTitle">Clients Checked In</h1>
+                  <Table responsive hover>
+                    <tbody>
+                      {this.props.pendingAccounts.map((account, index) => (
+                        <tr
+                          key={`button-to-accept-account-${index}`}
+                          onClick={() => this.props.dispatch({ type: 'UPDATE_PENDING_ACCOUNT', payload: { id: account.id, approved: true }})}
+                        >
+                          <td>
+                            {JSON.stringify(account)}
+                          </td>
+                        </tr>
+                      ))}
+                    </tbody>
+                  </Table>
+                </form>
+              </Card>
+            </Col>
           </Form.Row>
         </Container>
       </>
